requirements.txt
pandas<1.2.0
xlrd<2.O
Pillow==8.3.1
plotly==5.1.0
streamlit==0.84.1
openpyxl==3.0.6
